requirements.txt
# Add this in for tox, comment out for build
--extra-index-url https://download.pytorch.org/whl/cu117
torch==1.13.1
torchvision
torchaudio
xformers
torchdiffeq
torchsde
einops
open-clip-torch
transformers>=4.25.1
safetensors
pytorch_lightning
aiohttp
accelerate
pyyaml
pillow
loguru
clip-interrogator